*** empty log message ***

+++ examples/Connection/misc/Connection_Handler.cpp
@@ -155,7 +155,7 @@ Connection_Handler::handle_input (ACE_HANDLE)
       ACE_ERROR_RETURN ((LM_ERROR, 
 			 " (%P|%t) closing log daemon (fd = %d)\n", this->get_handle ()), -1);
     default:
-      if (buf[0] == EOF)
+      if (((int) buf[0]) == EOF)
 	ACE_ERROR_RETURN ((LM_ERROR, 
 			   " (%P|%t) closing log daemon (fd = %d)\n", this->get_handle ()), -1);
       else
